+++ libs/api/metadata-converter/src/lib/iso19139/codelists/role.mapper.ts
